requirements.txt
anthropic>=0.40.0
python-pptx>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
